AES: move test to capsules::test::
